+++ src/app/models/settings/GetSettingRequest.ts
@@ -25,9 +25,13 @@ export class GetSettingRequest {
 }
 
 const validator = GetSettingRequest.validator = JoiModelValidator.create({
-    slug: joi.string().regex(/^[0-9a-zA-z-]+$/).required().example('SettingSvc').example('setting-svc'),
-    ipAddress: joi.string().ip().required().example('127.0.0.1').example('192.168.10.23'),
-}, false, false)
+    schemaMapModel: {
+        slug: joi.string().regex(/^[0-9a-zA-z-]+$/).required().example('SettingSvc').example('setting-svc'),
+        ipAddress: joi.string().ip().required().example('127.0.0.1').example('192.168.10.23'),
+    },
+    isCompositePk: false,
+    requirePk: false,
+})
 
 validator.partial = function() {
     throw new NotImplementedException('This method is not supported. Use `whole` instead.')