+++ src/app/models/settings/SettingItem.ts
@@ -60,10 +60,15 @@ export class SettingItem {
 }
 
 SettingItem.validator = JoiModelValidator.create({
-    name: joi.string().token().required(),
-    dataType: joi.string().required().only(SettingItemDataType.String, SettingItemDataType.StringArray,
-        SettingItemDataType.Number, SettingItemDataType.NumberArray, SettingItemDataType.Boolean),
-    value: joi.string().allow('').required(),
-}, false, false)
+    schemaMapModel: {
+        name: joi.string().token().required(),
+        dataType: joi.string().required()
+            .only(SettingItemDataType.String, SettingItemDataType.StringArray,
+                SettingItemDataType.Number, SettingItemDataType.NumberArray, SettingItemDataType.Boolean),
+        value: joi.string().allow('').required(),
+    },
+    isCompositePk: false,
+    requirePk: false,
+})
 
 SettingItem.translator = new ModelAutoMapper(SettingItem, SettingItem.validator)