+++ src/app/translators/ModelAutoMapper.ts
@@ -82,23 +82,46 @@ export class ModelAutoMapper<T extends Object> {
      * Validates then converts an object to type <T>.
      * but ONLY properties with value are validated and copied.
      * Note that `required` validation is IGNORED.
-     * @param {any | any[]} source An object or array of objects to be translated.
+     * @param {object} source The object to be translated.
      *
      * @throws {ValidationError} If no `errorCallback` is provided.
      */
-    public partial(source: any | any[], options?: MappingOptions): Partial<T> | Partial<T>[] {
-        return this.tryTranslate('partial', source, options)
+    public partial(source: object, options?: MappingOptions): Partial<T> {
+        return this.tryTranslate('partial', source, options) as Partial<T>
+    }
+
+    /**
+     * Validates then converts a list of objects to type <T>.
+     * but ONLY properties with value are validated and copied.
+     * Note that `required` validation is IGNORED.
+     * @param {object[]} sources A list of objects to be translated.
+     *
+     * @throws {ValidationError} If no `errorCallback` is provided.
+     */
+    public partialMany(sources: object[], options?: MappingOptions): Partial<T>[] {
+        return this.tryTranslate('partial', sources, options) as Partial<T>[]
     }
 
     /**
      * Validates then converts an object to type <T>.
      * ALL properties are validated and copied regardless with or without value.
-     * @param {any | any[]} source An object or array of objects to be translated.
+     * @param {object} source The object to be translated.
      *
      * @throws {ValidationError} If no `errorCallback` is provided.
      */
-    public whole(source: any | any[], options?: MappingOptions): T | T[] {
-        return this.tryTranslate('whole', source, options)
+    public whole(source: object, options?: MappingOptions): T {
+        return this.tryTranslate('whole', source, options) as T
+    }
+
+    /**
+     * Validates then converts a list of objects to type <T>.
+     * ALL properties are validated and copied regardless with or without value.
+     * @param {object[]} sources The list of objects to be translated.
+     *
+     * @throws {ValidationError} If no `errorCallback` is provided.
+     */
+    public wholeMany(sources: object[], options?: MappingOptions): T[] {
+        return this.tryTranslate('whole', sources, options) as T[]
     }